refactor(core): rename config.py -> built_run.py
The module holds the `BuiltRun` bundle (and `DataConfig` / `EvalConfig` /
`RunInstance` / `EvalPGDConfig` / `TargetSites`), not a config schema — the
pydantic schema lives in the unchanged `configs.py`. The old name collided
conceptually with `configs.py` and the lab-side `experiments/config.py`.

Updated all `from param_decomp.config import` sites + CLAUDE.md refs.
`configs.py` (the schema) is untouched.
