+++ param_decomp_lab/clustering/scripts/run_worker.py
@@ -21,7 +21,7 @@ from pathlib import Path
 import jax.numpy as jnp
 import numpy as np
 
-from param_decomp.config import DataConfig
+from param_decomp.built_run import DataConfig
 from param_decomp.data import BatchSchedule, ShardServer, scan_shards
 from param_decomp.log import logger
 from param_decomp_lab.clustering.harvest_config import HarvestConfig
