+++ param_decomp_lab/tests/test_launch.py
@@ -1,5 +1,5 @@
 """The lab-side single-file config validator (`pd-lm`, torch venv). The runtime
-loader (`param_decomp.config`, jax venv) can't be imported here, so this exercises
+loader (`param_decomp.built_run`, jax venv) can't be imported here, so this exercises
 only the lab half: structural dispatch + the no-`run_id` precondition + stamping."""
 
 from pathlib import Path
